invoices: move update logic into separate file
